+++ apps/core/.vitepress/sidebar.ts
@@ -21,60 +21,47 @@ const SIDEBAR = {
         { text: "What Is BeaconKit?", link: "/learn/what-is-beaconkit" },
         { text: "Connect To Berachain", link: "/learn/connect-to-berachain" },
         { text: "How To Get $BERA", link: "/learn/how-to-get-bera" },
-      ],
-    },
-    {
-      text: "Berachain Testnet",
-      items: [
-        {
-          text: "Berachain Testnet V1 vs V2",
-          link: "/learn/testnet/berachain-testnet-v1-vs-v2",
-        },
+        { text: "Claim $BERA Airdrop", link: "/learn/claim-bera-airdrop" },
       ],
     },
     {
       text: "Proof-of-Liquidity",
       items: [
-        { text: "PoL Overview", link: "/learn/pol/" },
+        { text: "Overview", link: "/learn/pol/" },
         { text: "Participants", link: "/learn/pol/participants" },
+        { text: "Block Production/Emissions", link: "/learn/pol/bgtmath" },
         { text: "Reward Vaults", link: "/learn/pol/rewardvaults" },
         { text: "Incentives", link: "/learn/pol/incentives" },
         {
           text: "Tokens",
           items: [
-            { text: "$BERA", link: "/learn/pol/tokens/bera" },
-            { text: "$BGT", link: "/learn/pol/tokens/bgt" },
-            { text: "$HONEY", link: "/learn/pol/tokens/honey" },
+            { text: "Tokenomics", link: "/learn/pol/tokens/tokenomics" },
+            { text: "BERA", link: "/learn/pol/tokens/bera" },
+            { text: "BGT", link: "/learn/pol/tokens/bgt" },
+            { text: "HONEY", link: "/learn/pol/tokens/honey" },
           ],
         },
+        { text: "PoL FAQs", link: "/learn/pol/faqs" },
       ],
     },
     {
       text: "Governance",
       items: [
         { text: "Governance Overview", link: "/learn/governance/" },
-        {
-          text: "Reward Vault Whitelisting",
-          link: "/learn/governance/rewardvault",
-        },
+        { text: "Reward Vaults", link: "/learn/governance/rewardvault" },
       ],
-    },
-    {
-      text: "Guide",
-      items: [{ text: "PoL FAQs", link: "/learn/pol/faqs" }],
     },
     {
       text: "Native dApps",
       items: [
-        { text: "BEX", link: "/learn/dapps/bex" },
-        { text: "Bend", link: "/learn/dapps/bend" },
-        { text: "Berps", link: "/learn/dapps/berps" },
-        { text: "BGT Station", link: "/learn/dapps/bgt-station" },
-        { text: "Faucet", link: "/learn/dapps/faucet" },
+        { text: "BeraHub", link: "/learn/dapps/berahub" },
+        { text: "BeraSwap", link: "/learn/dapps/beraswap" },
         { text: "Honey Swap", link: "/learn/dapps/honey-swap" },
         {
-          text: `${constants.testnet.dapps.beratrail.name}`,
-          link: `/learn/dapps/beratrail`,
+          text: `${constants.mainnet.dapps.berascan.name}`,
+          link: `${constants.mainnet.dapps.berascan.url}`,
+          target: "_blank",
+          rel: "no-referrer",
         },
       ],
     },
@@ -93,7 +80,7 @@ const SIDEBAR = {
     {
       text: "Berachain For Developers",
       items: [
-        { text: "PoL Architecture Overview", link: "/developers/" },
+        { text: "Proof of Liquidity Overview", link: "/developers/" },
         {
           text: "Network Configurations",
           link: "/developers/network-configurations",
@@ -101,14 +88,8 @@ const SIDEBAR = {
         { text: "Deployed Contracts", link: "/developers/deployed-contracts" },
         { text: "Developer Tools", link: "/developers/developer-tools" },
         {
-          text: `${constants.testnet.dapps.beratrail.name}`,
-          link: `${constants.testnet.dapps.beratrail.url}`,
-          target: "_blank",
-          rel: "no-referrer",
-        },
-        {
-          text: `${constants.testnet.dapps.faucet.name}`,
-          link: `${constants.testnet.dapps.faucet.url}`,
+          text: `${constants.mainnet.dapps.berascan.name}`,
+          link: `${constants.mainnet.dapps.berascan.url}`,
           target: "_blank",
           rel: "no-referrer",
         },
@@ -119,11 +100,19 @@ const SIDEBAR = {
       items: [
         { text: "Build A Smart Contract", link: "/developers/quickstart/" },
         { text: "Build A Frontend", link: "/developers/quickstart/frontend" },
+        {
+          text: "Integrating with PoL",
+          link: "/developers/quickstart/pol-integration",
+        },
       ],
     },
     {
       text: "Developer Guides",
       items: [
+        {
+          text: "Mainnet Migration Guide",
+          link: "/developers/guides/migration-guide",
+        },
         {
           text: "Create HelloWorld Contract Using Hardhat",
           link: "/developers/guides/create-helloworld-contract-using-hardhat",
@@ -149,34 +138,36 @@ const SIDEBAR = {
     {
       text: "Contract References",
       items: [
-        { text: "Berachef", link: "/developers/contracts/berachef" },
+        { text: "BeaconDeposit", link: "/developers/contracts/beacondeposit" },
+        { text: "BeraChef", link: "/developers/contracts/berachef" },
         { text: "BGT Token", link: "/developers/contracts/bgt-token" },
         { text: "BGT Staker", link: "/developers/contracts/bgt-staker" },
         {
-          text: "Block Rewards Controller",
-          link: "/developers/contracts/block-rewards-controller",
+          text: "Block Reward Controller",
+          link: "/developers/contracts/block-reward-controller",
         },
         { text: "CREATE2", link: "/developers/contracts/create2" },
         { text: "Distributor", link: "/developers/contracts/distributor" },
         { text: "Fee Collector", link: "/developers/contracts/fee-collector" },
         { text: "Governance", link: "/developers/contracts/governance" },
         { text: "Honey Factory", link: "/developers/contracts/honey-factory" },
+        {
+          text: "Honey Factory Reader",
+          link: "/developers/contracts/honey-factory-reader",
+        },
         { text: "Honey Token", link: "/developers/contracts/honey-token" },
         { text: "Multicall3", link: "/developers/contracts/multicall3" },
         { text: "Permit2", link: "/developers/contracts/permit2" },
         {
-          text: "Rewards Vault",
-          link: "/developers/contracts/rewards-vault",
+          text: "Reward Vault",
+          link: "/developers/contracts/reward-vault",
         },
         {
-          text: "Rewards Vault Factory",
-          link: "/developers/contracts/rewards-vault-factory",
+          text: "Reward Vault Factory",
+          link: "/developers/contracts/reward-vault-factory",
         },
         { text: "TimeLock", link: "/developers/contracts/timelock" },
-        { text: "USDC Token", link: "/developers/contracts/usdc-token" },
         { text: "WBERA Token", link: "/developers/contracts/wbera-token" },
-        { text: "WBTC Token", link: "/developers/contracts/wbtc-token" },
-        { text: "WETH Token", link: "/developers/contracts/weth-token" },
       ],
     },
   ],
@@ -208,8 +199,24 @@ const SIDEBAR = {
           link: "/nodes/guides/kurtosis",
         },
         {
-          text: "Restoring Nodes from Snapshots",
-          link: "/nodes/guides/snapshots",
+          text: "Become A Validator Node",
+          link: "/nodes/guides/validator",
+        },
+        {
+          text: "Change Operator Address",
+          link: "/nodes/guides/operator-address",
+        },
+        {
+          text: "Manage Reward Allocations",
+          link: "/nodes/guides/reward-allocation",
+        },
+        {
+          text: "Distribute Block Rewards",
+          link: "/nodes/guides/distribute-block-rewards",
+        },
+        {
+          text: "Increase Validator $BERA Stake",
+          link: "/nodes/guides/increase-validator-bera-stake",
         },
       ],
     },