+++ apps/core/.vitepress/theme/index.ts
@@ -12,7 +12,6 @@ export default {
   extends: DefaultTheme,
   Layout: () => {
     return h(DefaultTheme.Layout, null, {
-      // https://vitepress.dev/guide/extending-default-theme#layout-slots
       "sidebar-nav-after": () => h(Fidget),
     });
   },