+++ apps/core/CHANGELOG.md
@@ -1,10 +1,22 @@
 # @berachain/core
 
+## 2.1.0
+
+### Minor Changes
+
+- Update contract interfaces and migration guide
+
+### Patch Changes
+
+- Updated dependencies
+  - @berachain/config@1.2.0
+
 ## 2.0.1
 
 ### Patch Changes
 
 - Update on validator system requirements
+- Update incentives logic
 
 ## 2.0.0